requirements.txt
aiofiles==0.4.0
alembic==1.0.8
aniso8601==3.0.2
appdirs==1.4.3
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
bcrypt==3.1.6
black==19.3b0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
dnslib==0.9.9
dnspython==1.16.0
email-validator==1.0.3
factory-boy==2.11.1
Faker==1.0.4
fastapi==0.9.1
graphene==2.1.3
graphql-core==2.1
graphql-relay==0.4.5
h11==0.8.1
httptools==0.0.13
idna==2.8
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.8
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==6.0.0
mypy==0.670
mypy-extensions==0.4.1
nejma==0.1.0
passlib==1.7.1
pluggy==0.9.0
promise==2.2.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pycparser==2.19
pydantic==0.21
PyJWT==1.7.1
pylint==2.3.1
pytest==4.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
python-multipart==0.0.5
PyYAML==5.1
requests==2.21.0
Rx==1.6.1
six==1.12.0
SQLAlchemy==1.3.1
starlette==0.11.1
text-unidecode==1.2
toml==0.10.0
typed-ast==1.3.1
ujson==1.35
urllib3==1.24.1
uvicorn==0.6.1
uvloop==0.12.2
websockets==7.0
wrapt==1.11.1
